Renaming file and refactoring class
